requirements.txt
msal==1.24.1
requests==2.31.0
aiohttp==3.9.1
python-dotenv==1.0.0
boto3==1.34.0